TrickWinner moved from Card to Round
