+++ src/test/java/ch/uzh/ifi/hase/soprafs22/screwyourneighborserver/api/CardIntegrationTest.java
@@ -1,0 +1,3 @@
+package ch.uzh.ifi.hase.soprafs22.screwyourneighborserver.api;
+
+public class CardIntegrationTest {}
